Clean up tagging makefile targets
Previously, when tagging a release, we would use `git push origin
--tags` to push the new tag to Github. This will do that, but will
also push all tags that you have locally. Since you'll likely already
have all the tags that the remote has, this results in a lot of noisy
output with lots of failed pushes. Really, all you want to do is push
the single new tag you made. So that's what we do now.

The `re-tag-release` target is also now just a composition of the new
`delete-release-tag` target and the existing `tag-release`
target. This gives a name for a new operation, and reduces duplication
of code.

Signed-off-by: Christopher Maier <cmaier@chef.io>