Add directory for default studio files
Rather than embedding file content in heredoc-style strings in a
script, we can create actual files for them.

This should be more flexible, and less susceptible to quoting errors.

We create a `defaults` directory, the internal structure of which is
meant to mimic a standard Linux filesystem. This directory is copied
into the package at build time, and as files are needed during studio
setup, they are copied out of this directory.

In addition to introducing this new infrastructure, this commit
transferrs the `/etc/inputrc` file over to this pattern, per
https://github.com/habitat-sh/habitat/pull/5658#pullrequestreview-210281676

Signed-off-by: Christopher Maier <cmaier@chef.io>