Remove uses of core's default_cache_key_path
See https://github.com/habitat-sh/habitat/issues/6314 for why.

This removes basically all the calls in the core hab CLI processing.
All of which are of the form:

    default_cache_key_path(Some(&*FS_ROOT))

and replaces them with cache_key_path_from_matches, which instead of
calling core's nonstandard env::var implementation, accesses values
which have been validated by clap. To facilitate this, and improve
discoverability, all the CLI commands which previously relied on
default_cache_key_path now have a --cache-key-path, which is documented
and overridable with the HAB_CACHE_KEY_PATH environment variable, but
in a safe way, where an empty-string value generates an informative
error.

Previously, empty-string values would be treated as specifying the
current working directory as the search root. This is no longer legal,
as it is too likely to be the result of an erroneous attempt to unset
the override environment variable. If this behavior is desired, the use
of the more explicit "." with either the HAB_CACHE_KEY_PATH env var or
--cache-key-path argument will achieve the same effect more clearly.

Additional uses of default_cache_key_path exist and will be addressed
in a subsequent commit.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>