Platform-specific UserInfo proof of concept
See https://github.com/habitat-sh/habitat/pull/6305#discussion_r267052477
for more context.

I don't think this is worth going forward with because adding `required`
types to launcher.proto shackles the launcher and supervisor together
in an undesirable way. But if we keep our protobuf-generated structs
with optional types, we are stuck with the decision between `unwrap`
during the infallible conversion from `generated::Spawn` -> `Spawn`,
or we can't really eliminate the `Option<_>`s in UserInfo.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>