Remove os_signal function from OsSignal trait
This function was only ever used in the Linux implementation, and only
inside this crate.

Furthermore, the existing `From<Signal> for i32` trait implementation
is unused, and essentially the same as the `os_signal` implementation
we had in the first place, but with less type-safety (since it was
using bare numbers and not referencing the various `libc` signal
constants).

So now we just implement `From<Signal> for SignalCode` and call it a
day.

Signed-off-by: Christopher Maier <cmaier@chef.io>