Work in progress
To fully eliminate default_cache_key_path, we need to have proper CLI
arguments for each command that needs a key cache path. That allows the
env var override to be checked appropriately.

Unfortunately, the studio (in the docker case) needs this value, but
the CLI parsing for the `hab studio …` commands is mostly deferred to
hab-studio.{sh,ps1}, meaning either we need to add the validation in
the scripts, do it ad-hoc in rust (not clap) in studio::enter::start,
or pull the CLI parsing into our standard clap-based approach. The last
option seems the best, so this is a start down that road.

Still to do are implementing the rest of the `hab studio …` subcommands
in clap and possibly further refactoring of studio::enter::start to
completely remove the `args: &[OsString]` argument. Finally, it likely
makes sense to completely remove the argument parsing code from the
hab-studio scripts once this is in place, but that approach needs more
thought first.

Refactoring out the rest of the ad-hoc arg parsing in exec_subcommand_if_called
is probably a good idea, but that can be in a follow-up PR.

See https://github.com/habitat-sh/habitat/issues/6314

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>