Further unification and simplification
* Unify shutdown checking logic.

  The implementation on Windows and Unix is effectively the same, so
  we've moved things around so now they are *exactly* the same.

  This also meant that there's not much benefit at all to having a
  separate Windows module. So now it's gone :)

* Move `SignalEvent` into `unix` module.

  Since it's only applicable on Unix, it makes more sense to push it
  further down into the platform-specific code.

Since there's still a good amount of platform-specific code for Unix,
it makes sense to continue to keep it in the `unix` module (otherwise
we'd have a lot more conditional compliation in the main module, which
can be confusing.

This _does_ mean that the implementation of `init` is "split" across
two modules, and that the Unix implementation of it has to point to
the static `SHUTDOWN` boolean in the root module, but on balance, it
seems like a decent compromise.

Signed-off-by: Christopher Maier <cmaier@chef.io>