Properly pass CLI args to `hab origin key import`
The argument matches for the `key` supercommand rather than the
`import` subcommand were passed to sub_origin_key_import, causing the
unwrap of the required CACHE_KEY_PATH argument to fail.

Fix this, plus change the unwrap to an expect (this is best even when
you don't expect there's any way to fail) and move the read of stdin
to after the arg parsing so this can be more easily triggered.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>