Consolidate feature flag operations.
Previously, feature flags were defined in both the `hab` and
`habitat_sup` crates. This was problematic because it required far too
much setup in order to create a feature flag.

Additionally, if you wanted to create a feature flag that is mainly
used in the Supervisor, but conditionally adds additional CLI options,
you were out of luck. Because of underlying code, as well as the fact
`hab` and `hab-sup` are entirely separate binaries, the feature flags
enabled in `hab` wouldn't necessarily be enabled in `hab-sup`.

This refactoring moves *all* feature flags into the common
crate. Additionally, we dispense with the use of the `features` crate,
in favor of using the underlying `bitflags` abstraction directly.

This means that we no longer rely on recording which feature flags are
currently enabled in global state, but in immutable data that we
explicitly thread through function calls.

This also means that functionality that differs based on feature flag
state can be more easily unit tested (otherwise, we have to contend
with the sharing of global state, which is always tricky).

Signed-off-by: Christopher Maier <cmaier@chef.io>