Fix up the `curl`-based package upload override
The previous workaround wasn't working for a handful of reasons. The
main one appears to be that we were originally using `http` instead of
`https`, but weren't properly following the redirect. Switching to
`https` solves this.

Another problem was the lack of channel in the original promotion
`curl` command.

This commit also introduces the use of the `--fail` option, in order
to convert 4xx and 5xx responses into errors which will halt the
build.

Additionally, when using raw `curl`, we must manually create the
channel we are promoting packages into. This command does *not* use
`--fail`, because we want to continue on if the channel already
exists.

I also took the liberty of using the long option names for everything,
in the interest of readability.

Signed-off-by: Christopher Maier <cmaier@chef.io>