Add promotion workaround at final promotion step
First, since we're doing this promotion workaround all over the place,
we might as well extract it into a standalone function.

Second, in order to use it in the final stage of the pipeline where we
promote from the release candidate channel into the `builder-live` and
`stable` channels, we need to know the target. Unfortunately, we can't
yet get that from Builder without already knowing it. As such, the
most expedient thing to do is to use Buildkite's metadata
functionality to track it.

Each time we promote something into our release candidate channel,
we'll create a metadata key with the identifier and target of that
package. Then, at the end when we just have identifiers of the
packages in the release candidate channel, we loop through each of our
targets for each identifier. If we've got a corresponding metadata
key, then that means that package is for that particular
target. Knowing that, we should be able to make the `curl`-based
promotion call.

Signed-off-by: Christopher Maier <cmaier@chef.io>