Be explicit when dealing with feature flags
Here, we dispense with the use of the `features` crate, in favor of
using the underlying `bitflags` abstraction directly.

This means that we no longer rely on recording which feature flags are
currently enabled in global state, but in immutable data that we
explicitly thread through function calls.

This also means that functionality that differs based on feature flag
state can be more easily unit tested (otherwise, we have to contend
with the sharing of global state, which is always tricky).

Signed-off-by: Christopher Maier <cmaier@chef.io>