Customize service shutdown logic (Part 1)
Previously, Habitat would send a `TERM` signal (on Linux; Ctrl-C on
Windows) to a service to shut it down. If, after 8 seconds, the
service was still alive, Habitat would forcefully kill the
process. This is not ideal, since some services may need to respond to
a different signal, or take longer than 8 seconds to complete a
shutdown process.

Here, we add a new feature flag (available with the environment
variable `HAB_FEAT_CONFIGURE_SHUTDOWN`) that allows users to customize
how the Supervisor stops a service from the `hab svc stop` and `hab
svc unload` commands. With this feature enabled, those commands gain a
`--shutdown-timeout` option, and (on Linux) a `--shutdown-signal`
option.

These parameters are then plumbed all the way through to modify the
shutdown process in the desired way. If the feature is not enabled, or
one or more of the options are not used, then the currently-hardcoded
values are used as defaults.

An example of how to use this feature:

```sh
export HAB_FEAT_CONFIGURE_SHUTDOWN=1
sudo hab svc stop my/nifty-service \
     --shutdown-signal=INT \
     --shutdown-timeout=20
```

This would send a `SIGINT` signal to `my/nifty-service` and wait 20
seconds for it to terminate before sending a `SIGKILL`. A Windows
invocation would look similar, but without the `--shutdown-signal`
option.

Importantly, this is not yet a complete feature. In particular,
services shutting down for automatic updates do not have access to any
customized shutdown parameters. To do this, we will need to make it
possible to set these parameters at service load-time, storing them in
the service's spec file for access later. Additionally, we will likely
want to modify service plans to add the ability for plan authors to
specify these parameters.

Internally, we have modified the communication pattern between the
control gateway (that receives user commands) and the main Supervisor
logic. Previously, when stopping or unloading a service, the control
gateway functions would either write to or delete the appropriate
service spec file; after some delay, the Supervisor would notice a
file change event and take the appropriate action. That is, the
communication was indirect, mediated by the filesystem. This made it
impossible to convey user input (namely, the customized shutdown
parameters) to the Supervisor.

Now, `hab svc stop` and `hab svc unload` no longer operate on the
file, but rather send a message through a channel into the Supervisor
itself. As the Supervisor executes its main loop, it will pull
messages from this channel and take the appropriate actions. The spec
file is still ultimately modified or deleted, but the Supervisor does
it now.

Long term, this provides a more direct pattern of interaction that
will eliminate the need for filesystem monitoring of spec files
altogether, and may make things like synchronous load and unload
possible.

The messages transmitted across this channel are of the new
`SupervisorAction` type. Currently this enum has variants for service
unload and service stop, but will grow to encompass additional
actions.

As part of these new supervisor actions, we also create a new
`ShutdownSpec` type to encapsulate the shutdown signal and timeout
parameters. We do this encapsulation to better deal with
platform-specific behavioral differences between Unix and Windows;
namely, the fact that signals don't exist on Windows. Previous code
dealing with signals was not explicit about this, which made the
resulting code more complex than it needed to be.

Additional Refactorings
=======================

In the course of modifying the control gateway to use this new
interaction pattern, several refactorings were performed. They are, in
no particular order:

* Extract `CtlAcceptor` and `CtlHandler` to separate modules

These futures were previously defined in the manager module. They are
now defined in submodules of the control gateway, which is more
appropriate.

* Removed `Box` in future-returning signatures, in favor of `impl Future`

This is more idiomatic, now that `impl` return values exist.

* Variable renaming

With more channels in play, simple names like `rx` and `tx` no longer
cut it. Now, things are named more descriptively.

* Modify `use` statements

Specifically, we now use `habitat_sup_protocol::codec` types
explicitly, rather than using a wildcard. We also use
`tokio_codec::Decoder` instead of `tokio_codec::Framed` for more
flexibility (the former is a trait; the latter is a struct) and
readability.

* General rearrangements of future-creation chains for clarity and readability

* Remove unnecessary `return` calls

Clippy doesn't like them, and neither do I.

* Extract `command_from_message` and `start_timer` functions

This makes it much easier to understand the logic flow of the
`SrvHandler` future.

* Move spec-related functions into `ManagerConfig`

With some command logic living in the Supervisor, while other logic
remains in the control gateway, we need to share these previously
private functions. Since they all use a `ManagerConfig` argument, they
have been moved to methods on the `ManagerConfig` type.

Signed-off-by: Christopher Maier <cmaier@chef.io>