Make --binlink take an value (or env var) for `hab pkg install`
Resolves https://github.com/habitat-sh/habitat/issues/6418

Commit 6232c6728f29f825b4324b81ff7c77b7eec95d79 removed
`binlink_dest_dir_from_matches`, which applied the value in the
HAB_BINLINK_DIR env var to choose the destination when the `--binlink`
flag was provided. This resulted in a bug where the binlinks weren't
in the expected location in a new studio.

This commit changes `--binlink` from a flag to a option that takes a
value. That value has platform-specific defaults, allowing the existing
use case of `--binlink` without a value to work with either the default,
or with a value supplied by the HAB_BINLINK_DIR env var.

One notable change is that since `--binlink` now takes a value, it must
be placed *after* the `<PKG_IDENT_OR_ARTIFACT>` positional argument if
the default value is desired. Otherwise, something like
`hab pkg install --binlink core/redis` would interpret `core/redis` as
the binlink path. Fortunately, this would be caught by clap since the
command appears to lack the required `<PKG_IDENT_OR_ARTIFACT>`
positional argument, yielding the error:
```
error: The following required arguments were not provided:
    <PKG_IDENT_OR_ARTIFACT>...

USAGE:
    hab pkg install <PKG_IDENT_OR_ARTIFACT>... --binlink <BINLINK> --channel <CHANNEL>
```

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>