Add habitat_common::sync::Lock type for recursive read locking
Lock provides a wrapper around a parking_lot `RwLock` so we can add in
safe recursive read locking while adding the ability to transparently
replace the RwLock with a Mutex which makes it easier to track down
recursive locking.

This adds features `lock_as_rwlock` and `lock_as_mutex` to the
habitat_common crate, as well as the `habitat_butterfly` and
`habitat_butterfly_test` crates which depend upon it. If neither
is enabled, `Lock` uses `RwLock`, though this isn't a default feature,
otherwise it would require `--no-default-feature` on the CLI any time
we wanted to use `lock_as_mutex`.

This also enables `lock_as_mutex` for the butterfly and sup CI tests.

See https://github.com/habitat-sh/habitat/issues/6435

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>