Streamline GatewayState struct
In preparation for some additional refactorings, this commit tightens
up the definition and role of the `GatewayState` struct. Now, it
consists solely of data that is ultimately served over the
Supervisor's HTTP gateway.

The gateway's optional authentication token is better understood as a
property of the gateway server itself. Thus, `GatewayAuthToken` is
moved into the `http_gateway` module (and renamed
`GatewayAuthenticationToken` to remove ambiguity). It is no longer
part of `GatewayState`, but is now part of the gateway's `AppState`
struct.

Signed-off-by: Christopher Maier <cmaier@chef.io>