[PR Feedback] Let Windows clients send signals to Unix services
Previously, all the Signal-related types were strictly Unix-only, in
an effort to help clarify what internal code truly applied to Unix
vs. Windows. I was a bit overzealous in carrying this all the way out
to the CLI.

This change makes it possible for Windows clients to send customized
shutdown signals to services running on Unix-based Supervisors.

Signed-off-by: Christopher Maier <cmaier@chef.io>