Update the default allocator to jemalloc.
Part of this is just to be consistent with hab-sup, which uses jemalloc
because it allows the gathering of metrics with regards to allocations.
The other part is to try and solve some strange hanging issues where
'hab pkg path' will peg the CPU at 100% and never return. Attaching gdb
to a hung process and inspecting stack frames shows malloc as the
culprit, so this is an attempt to fix that bug.

Signed-off-by: Josh Black <raskchanky@gmail.com>