Add cancelable_future
This future provides a convenient way to retain control over a
potentially long-running future that you might want to cancel.

To accomplish this, we create a oneshot channel to act as the means of
interacting with the long-running future. We then select across the
long-running future and the receiving end of the channel. Thus, the
long-running future can be short-circuited when the sending end of the
channel is either dropped or explicitly sends a message. We wrap the
sender in a `FutureHandle` to lock down the interface.

Signed-off-by: Christopher Maier <cmaier@chef.io>