Revise deadlock detection approach
Deadlock detection doesn't work if one of the blocked threads is
calling a try lock variant (including one with a timeout), so there's
no way we can reasonably recover from a detected deadlock. Instead, we
start one deadlock detector thread if the feature is enabled and abort
the whole process if one is discovered.

Also, add separate CI steps for running butterfly tests with the
different lock features, but don't fail the entire CI run if there are
failures in the `lock_as_mutex` version.

Finally, add `--quiet` to the `cargo test` invocations to suppress the
largely irrelevant compilation output and get straight to the test
run output.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>