Add some automated networked tests
This introduces some `docker-compose` based testing scenarios for
exercising various Habitat behaviors in small networks. Read the
comments in the code for further details.

Right now, the tests are written in Bash, with a very crude test
runner also written in Bash. As a result, they only really run on
Linux right now. However, we are investigating transitioning to
Powershell, as that can run on all our platforms, which means we could
get these same tests running on Windows without a lot of extra work.

One big downside of the current implementation is that it's difficult
to run the tests locally on code you're currently working on (instead,
you provide the name of a channel from which to pull Habitat
artifacts). More work needs to be done on this to ensure that `cargo
build`-built binaries can be self-contained enough to drop into a
container and still work. The benefit of having tests that run in CI,
however, is high, so we're getting this out now.

Signed-off-by: Christopher Maier <cmaier@chef.io>