Allow building of docker containers with license changes.
Concretely, this passes HAB_LICENSE=accept-no-persist to the docker
build process. This allows the container to be built but it specifically
does NOT persist the license acceptance file inside the container. This
means that when the container is run, it will fail unless
HAB_LICENSE=accept-no-persist is passed into the container, either by
the scheduler or via 'docker run'.

Signed-off-by: Josh Black <raskchanky@gmail.com>