Encapsulate update strategy with channel for events
Currently, a channel field is set in a service spec whether or not
that service is set to update. If the service is not updating, this
channel is meaningless; it indicates what channel the service was in
when it was initially downloaded, but only if it was downloaded in the
course of the `hab svc load` call. Additionally, a channel is not an
intrinsic feature of a package, but is situational and somewhat
dependent on the time the call was made.

For event purposes, we really only care about the channel if we are
also updating the service. To that end, this commit introduces a new
protobuf type to encapsulate both aspects of this relationship: the
variant of the update strategy, and the channel from which to pull
updates.

Importantly, there is no `UpdateStrategy::None` variant for the
protobuf type that corresponds to the existing domain type. This is
instead essentially converted into an `Option::None`, which is more
idiomatic in Rust.

This type currently exists only in the event protbuf layer, but it is
hoped that it could be a guide for a similar and more in-depth
refactoring to the broader Supervisor.

Signed-off-by: Christopher Maier <cmaier@chef.io>