Backing out cfg_attr changes for now
Looks like `path` attributes are invisible to rustfmt:

- https://github.com/rust-lang/rustfmt/issues/1208
- https://github.com/rust-lang/rustfmt/issues/2407

Ideally, it would be nice to not use the `path` attribute at all, and
we can actually formulate this in a way that doesn't use `path`, but
it seems to involve a lot more repetition. It's not at all clear that
that's a win here, particularly given that we're only going to have
two implementations for a short period of time.

Signed-off-by: Christopher Maier <cmaier@chef.io>