Add thread heartbeats to log errors for likely deadlocks
The 0.80.0 release incident showed that deadlock scenarios can be very
confusing to debug in production, so this adds calls to the various
threads which run a continuous work loop to indicate that they are
still processing and adds a new thread which checks the heartbeats and
logs an error indicating the thread which is likely deadlocked if one
does not occur for too long a period.

`ThreadAliveThreshold` (how long a heartbeat must go missing before an
error is logged, default 5 minutes) and `ThreadAliveCheckDelay` (how
frequently the checker thread looks for updated heartbeats, default
1 minute) can be configured with the `HAB_THREAD_ALIVE_THRESHOLD_SECS`
and `HAB_THREAD_ALIVE_THRESHOLD_SECS` environment variables.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>