Convert ServiceUpdatePeriod to use env_config_duration macro
Also, improve the macro to specify the associated function that should
be applied to the value in the environment variable to create the
Duration (for example: from_secs).

Other cleanup includes moving MIN_ALLOWED_PERIOD from a lazy_static
into an associated const (possible now that we're using the stdlib
Duration, which has const fn initializers) and removing trait
implementations which are used only in tests in favor of a bit more
direct access in test code.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>