Downgrade clippy::cyclomatic_complexity from error to warning
cyclomatic_complexity was renamed to cognitive_complexity in rust 1.35,
but because our CI defaults to `stable`, this resulted in a situation
where linux and windows run CI against different versions, leading to a
mutually incompatible situation for this lint. This is a temporary fix
until we can fully address the toolchain versioning issue in CI.

See https://github.com/habitat-sh/habitat/pull/6593 for more.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>