Add a rumor expiration mechanism.
There are a few parts to adding a rumor expiration mechanism. First, the
rumors themselves need an expiration date. This was added through a new
subtrait, called Transient, since not all rumors expire. Departure
rumors in particular should never expire because doing so would break
the mechanism that 'hab sup depart' was designed for.

Expired rumors are purged in the same background thread that handles
transitioning Members from Suspect -> Confirmed and Confirmed ->
Departed. The purge interval defaults to 60 seconds but is configurable
via the HAB_EXPIRE_THREAD_PURGE_SECS environment variable. The interval
that the thread checks at is now also configurable via the
HAB_EXPIRE_THREAD_SLEEP_MS environment variable and defaults to 500ms.

Rumors that are still relevant are refreshed periodically by the
processes that spawned them. Any rumors that are not refreshed within
the expiration window transition to expired and are purged.

The rumor heat mechanism was also removed.

The rumor heat mechanism described in the Xerox paper that our rumor
implementation is based on uses the heat to determine how many times
a rumor has been sent to the network as a whole. Our implementation
didn't do that. Instead, it used the heat to determine how many times
a rumor had been sent to each member of the network.

It turns out that's not necessary because we're using zmq as
a transport for the rumors, which uses TCP under the hood, so delivering
a rumor once is sufficient.

Now we just send out any rumors we have that aren't expired, which is
closer to an active anti-entropy approach than rumor mongering. Note that
this is still sub-optimal, because we're sending rumors that our
neighbors could potentially already have. In theory, network traffic
will be somewhat reduced, given that there will be less rumors to send,
and we will only be sending rumors once.

Signed-off-by: Josh Black <raskchanky@gmail.com>