Add rst-reader back in.
This reverts 94156b3620683ff97802526183ba68c93bc1afd4 as well as makes
various improvements on top of that reversion. In particular, DatFile
now has two separate initializers: one for the butterfly Server, which
will need to create the file if it doesn't already exist, and one for
rst-reader (or similar utilities) that should have read-only access, and
shouldn't be creating this file if it doesn't exist.

Signed-off-by: Josh Black <raskchanky@gmail.com>