Annotate RumorStore::remove locking and remove ::with_keys
RumorStore::with_keys is no longer necessary, since the addition of
Iterable guard allows more ergonomic for loops to be used instead.
In general:

    rs.with_keys(|(service_group, rumors)| { ... });

can be replaced with

	for (service_group, rumors) in rs.lock_rsr().iter() { ... }

Additionally, the flexibility of Iterator, allows for the use of all
the iterator adapters, so paradigms like:

    let mut v = vec![];
    for (service_group, rumors) in rs.lock_rsr().iter() { ... }
    v

can be replaced with more functional equivalents:

    rs.lock_rsr().iter().filter_map(|(service_group, rumors)| { ... })

This change just does the former, to minimize the potential for
changing behavior.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>