DatFile can't own the BufReader after all.
It turns out that when DatFile owns the BufReader, it keeps the file
open and prevents AtomicWriter from renaming the file on Windows.
Previously, a single BufReader was created and passed around between all
the different functions and that is the solution I've gone back to here
as well.

At first, I thought it should be possible to just create a BufReader
whenever necessary, but it turns out that BufReader keeps track of where
it is in the file, and our current implementation is dependent on
reading the rumors in the same order that they're written in.

Since we store the offsets for all the different rumors, in theory it
should be possible to use that to calculate where to go and how much to
read, but in practice, it proved trickier to implement correctly than
I had originally anticipated.

In the interest of time and sanity, we now pass the same BufReader
around, like we used to. Eventually, when someone has the time and
energy, this whole module should be overhauled.

Signed-off-by: Josh Black <raskchanky@gmail.com>