Remove thread-marking from SrvHandler future
This loop is actually a loop over the processing of a single command
request. Whenever a new request comes in, the futures reactor polls
this future again to process it. Thus, the two loops are totally
separate, and this thread-marking serves to tell us how long ago the
last command processing was. As a result, leaving it in can give the
erroneous impression of potentially deadlocked code in this module.

Signed-off-by: Christopher Maier <cmaier@chef.io>