Stop health check futures when upgrading the Supervisor
When we shut down the Supervisor normally, we stop the long-running
control gateway future, and then spawn a bunch of service shutdown
futures. Those aren't "infinite" futures, so they'll eventually be
driven to completion. Then, we just wait for all the futures on the
reactor to finish and we shut down.

When we shut down for a Supervisor upgrade, however, we _don't_ shut
down the services. Since their health checks now run in an infinite
future, these futures will effectively prevent the Supervisor from
shutting down.

This commit introduces a "detatch" method for services, which is now
called in the Supervisor upgrade scenario. It currently just stops
health checks, but can be augmented to do more as we transition more
functionality over to asynchronous futures.

Similarly, the existing "reattach" logic is now extracted into a
"reattach" method, which will now restart health checking.

Add minor tweaks to the existing Expect test that should have caught
this regression.

Fixes #6712

Signed-off-by: Christopher Maier <cmaier@chef.io>