Add liveliness checker regression test
This should cover the issue fixed in https://github.com/habitat-sh/habitat/pull/6724

This commit is expected to fail CI in order to validate the test.
The behavior will be re-fixed in a subsequent commit.

Additionally, improve the logging for the liveliness checker and switch
its locking to parking_lot from the stdlib Mutex.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>