Fix liveliness checking locking
Taking the lock *after* delaying makes more sense because the checker
is unlikely to have any interesting to report immediately after being
spawned. Plus, this avoid the need for an extra scope to ensure the
lock is not held during the delay.

The spawn_thread_alive_checker_allows_timely_heartbeats test validates
that this approach resolves the issue.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>