Be more explicit about connection refused errors.
We shouldn't have been casting every IO error as a failure to connect to
a supervisor. There are a wide variety of IO errors and only one of them
means we couldn't connect to a supervisor. Now we trap that error
specifically, and display the appropriate wall of text. Other IO errors
are displayed in a more generic fashion.

Signed-off-by: Josh Black <raskchanky@gmail.com>