hab: don't start analytics thread if analytics is disabled
This will introduce a few calls to stat(2) into the startup of Habitat
in exchange for not starting the analytics thread.

The actual motivation here is that this thread is involved in the
conditions that appear to be triggering:

https://github.com/habitat-sh/habitat/issues/6252

While we should fix that bug, not starting a thread we don't need
seems like an easy way to lower the impact of the bug quickly.

Typically, I can reproduce 6252 in between 10k-250k invocations of
hab. With this fix in place, I've been able to run 5M+ invocations of
hab without issue.

Signed-off-by: Steven Danna <steve@chef.io>