Transition other install_latest_hab_binary uses to curlbash_hab
When promoting packages and deleting channels, we don't really care
that we're using the version of `hab` that we're building or testing
in the pipeline. Rather, we just care that there's a `hab` binary
present to do basic operations with.

For that, we really just need to use `curlbash_hab`. Also note that we
don't care for the `hab_binary` global variable that
`install_latest_hab_binary` set. Since the "curlbash" method adds a
link to `/bin/hab`, we can just reference the binary by name directly.

Signed-off-by: Christopher Maier <cmaier@chef.io>