Use HAB_BLDR_URL directly
It appears that adding `HAB_BLDR_URL=$ACCEPTANCE_HAB_BLDR_URL` in a
step definition doesn't properly do the interpolation we want, and
instead sets `HAB_BLDR_URL` to the empty string.

I suspect the interpolation is being done when the pipeline definition
is being uploaded to Buildkite, which is why the value is not set.

Signed-off-by: Christopher Maier <cmaier@chef.io>