Explicitly manage the version of Nightly Rust in tests
Some of our tests use nightly Rust to help test for specific
things. The version of Rust used was simply piggybacking on the
version of nightly Rust we use for `rustfmt`, rather than being
specifically managed.

Previously, this wasn't an issue, but with the latest nightly Rust
that works for `rustfmt`, we apparently have a bug around cargo
dealing with crates that have symlinks in them (possibly related to
https://github.com/rust-lang/cargo/issues/5664).

This commit introduces a `RUST_NIGHTLY_VERSION` file to explicitly
manage which version we're using in tests. For now, it also pins back
to the last known good nightly version for testing purposes.

Signed-off-by: Christopher Maier <cmaier@chef.io>