mention earlier there's a debug option for PS plans
I'd been tripped up for weeks on the note saying there is no "attach"
command for PowerShell plans. I stopped reading there and totally missed
the section below it about "Set-PSBreakPoint". Here's an documentation
offering that might clue future readers into the existence of a
debugging option for PowerShell by mentioning both in the section's
intro paragraph and naming the sub-sections according to the plan-type.

Signed-off-by: Robb Kidd <robb@thekidds.org>