Always build the Launcher and Windows Service packages
Previously, we would only build Launcher and Windows Service packages
only if a user indicated manually that they should be built. As we
move toward a continuous release model, this manual step becomes
cumbersome. In any event, an engineer will _still_ have to evaluate
what has changed in the Launcher and Windows Service since the last
release in order to determine whether end-users will need to update
their Launcher packages. As a result, we're not really removing a
human quality gate here; we're just simplifying the pipeline.

For reviewers: the only concrete change here is making the existing
build steps run all the time; the content of the steps hasn't changed.

Signed-off-by: Christopher Maier <cmaier@chef.io>