Remove the trace_it! macro and associated functionality.
Ultimately, this feature isn't being used and its utility is limited to
just a couple of supervisors. Any more than that, and the output becomes
too much to make sense of. As part of cleaning up our deadlocks, we're
removing this. We might revisit something along these lines in the
future, but better suited to larger networks.

Signed-off-by: Josh Black <raskchanky@gmail.com>