Fix launcher_exits_on_supervisor_startup_failure test
Because of how TESTING_FS_ROOT works, this test was succeeding when it
wasn't really testing what it was supposed to.

This gets around that by:

a) increasing the retry count so we have enough time to install new packages, and
b) pre-installing the launcher

I don't love this solution, but it works until we can remove
TESTING_FS_ROOT entirely and can rethink the test.

Signed-off-by: Christopher Maier <cmaier@chef.io>