Change ManagerState::services to use sync::ManagerServices
This new type encapsulates the locking behavior (and switches from the
stdlib RwLock to the parking_lot-backed abstract Lock), making the lock
ordering simpler to document and enforce as well as moving more of the
complexity away from the consumers. For example:

	for (_ident, svc) in self.state
	                         .services
	                         .write()
	                         .expect("Services lock is poisoned!")
	                         .drain()
becomes:

	for svc in self.state.services.lock_msw().drain_services()

Also, habitat_butterfly::Server::suitability_lookup changes from an
Arc<Box<dyn Suitability>> to Arc<dyn Suitability>, and the need for a
separate SuitabilityLookup struct is eliminated by implementing the
Suitability trait on ManagerServices directly.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>