Remove automated Homebrew release script
There have been internal release engineering changes lately that have
deprecated some of the functionality this script depended on. Rather
than jump through hoops to fix the script, we are temporarily removing
it.

We are currently in the process of migrating the release process over
to the supported framework anyway, so this functionality will be
reintroduced soon.

Signed-off-by: Christopher Maier <cmaier@chef.io>