`is_alive` returns false for PID 0 on Unix
For as-yet-undetermined reasons, we have several reports of PID files
containing the PID of "0". This ends up causing problems, particularly
when the PID files are for services we are monitoring. When checking
to see if a service's PID is still alive, we use the `libc::kill`
call. If the supplied PID is 0, then the call basically amounts to "is
the current process alive?", which is always true. Unfortunately, this
leads the Supervisor to think it is supervising itself!

It gets worse if we think a service has a PID of 0 when it's time to
shut that service down. Here again we use `libc::kill`, but this time,
the Supervisor will end up terminating itself!

In this commit, we begin to lay the groundwork for unraveling this
mystery, while also providing some degree of safety for users until
the bug is ultimately squashed. Whenever we call `is_alive` with a PID
of 0, we will return `false`, since this will ultimately cause the
Supervisor to remove the offending PID file and try again.

This is not a complete solution, however.

PS: This behavior existed already on Windows, so we only add a few
explanatory comments there.

Signed-off-by: Christopher Maier <cmaier@chef.io>