Treat PID files containing "0" as corrupt
As a possible prelude to the creation of a PID file type to
encapsulate this behavior better, here we treat a PID file that
contains a "0" to be corrupt.

Specifically, we deal with service PID files as well as the Launcher /
Supervisor lock file.

This should prevent the system from being affected by such PID files.

It does not yet explain *how* those files came to contain a "0", but
it should make the system incrementally safer until the mystery is
solved.

Signed-off-by: Christopher Maier <cmaier@chef.io>