Add more logging around pid file shenanigans
This commit adds a number of new log messages, as well as modifies a
few preexisting ones, in an effort to shed more light on how a PID of
0 could legitimately occur.

Some of these may be obviated by changes in the preceeding commits,
but I'd rather err on the side of caution.

All messages have a target of "pidfile_tracing", allowing them to be
specifically singled out.

For instance, setting `RUST_LOG=pidfile_tracing=trace` will ensure
that all these log messages are present in logging output. Further
control can be gained by configuring the Supervisor's logging
configuration file (see
https://www.habitat.sh/docs/reference/#sup-log-configuration). However,
the environment variable will be needed for the Launcher's logging,
since it does not yet respond to that configuration file.

Signed-off-by: Christopher Maier <cmaier@chef.io>