Reply back to the Supervisor if it sends an unknown message type
Previously, if the Launcher received a message that it could parse as
a real launcher protocol message (i.e., `NetTxn`), but a message that
it didn't know what to do with, it would just drop it on the floor and
do nothing.

If we were to add a new message to the protocol, and a Supervisor that
used that message were interacting with an older Launcher that didn't
recognize that message, the Supervisor would actually hang, because
its interactions with the Launcher currently use a blocking `recv`
call to wait for a reply (which, in this case, will never come). In my
testing, I also discovered that attempting to shut things down
by, (e.g., sending SIGINT to the Launcher / pressing Ctrl-C) could end
up mutually deadlocking both Launcher and Supervisor.

There are more things that can be done to make this interaction more
robust, but the very least we can do is reply back to a Supervisor
that it's sending something we don't know what to do with.

Signed-off-by: Christopher Maier <cmaier@chef.io>