Switch from VecDeque to AtomicBools
This change moves us from using a VecDequeue under a Mutex for
tracking signals to signal-specific AtomicBools, replacing the generic
`check_for_signal()` function with two more-specific
`pending_sighup()` and `pending_sigchld()` functions.

This has two advantages:

1) No mutexes or memory allocations in our signal handling code. This
   matters less since we have also moved to synchronous handling of
   signals, bug gives us some flexibility to change our approach more
   safely in the future.

2) Clearer alignment between this interface and the application
   behavior.

The old check_for_signal() API looks like it intended to provide the
groundwork for a variety of signal-based behavior; however, in
reality, we are handling a small number of signals. The generic nature
of that interface and the fact that the code is shared between two
processes was obscuring the fact that we actually have a small amount
of signal-related behavior across the two services:

*Linux Behavior*

| signal    | hab-launch behavior | hab-sup behavior     |
|-----------+---------------------+----------------------|
| SIGINT    | Graceful shutdown   | Handled but ignored  |
| SIGTERM   | Graceful shutdown   | Handled but ignored  |
| SIGHUP    | Send to hab-sup     | Shutdown for restart |
| SIGCHLD   | Reap zombies        | Handled but ignored  |
| SIGQUIT   | Handled but ignored | Handled but ignored  |
| SIGALRM   | Handled but ignored | Handled but ignored  |
| SIGUSR1   | Handled but ignored | Handled but ignored  |
| SIGUSR2   | Handled but ignored | Handled but ignored  |
| All other | Default disposition | Default disposition  |

*Windows Behavior*

Windows uses a library which appears to handle the CTRL_C_EVENT. The
set of changes should not change the behavior here, although we are
now installing the handler earlier in the startup of the relevant
applications.

This more explicit interface has the disadvantage of not being
well-suited to a larger variety of signal-based behavior. However, it
seems to me that we can always implement something smarter when that
actually becomes an issue.

Signed-off-by: Steven Danna <steve@chef.io>