Reorder pipeline stages for maximum parallelism
While the `windows-service` package must be built after the Windows
Launcher package, there is nothing that prevents Launcher and
Supervisor packages from being built in parallel, since they are
mutually independent on a Habitat dependency level.

Similarly, none of the exporters depend directly on either the
Launcher or the Supervisor, so they can all be built at the same time
as those two packages.

Now, we build the Windows Service toward the end, alongside the
`core/hab-pkg-cfize` exporter, which itself has a hard dependency on
`core/hab-pkg-export-docker`.

By reordering these and building more packages in parallel, we should
be able to get through the entire pipeline in less overall time.

Signed-off-by: Christopher Maier <cmaier@chef.io>