Fully qualified package names download irrespective of channel
Fix for hab 7039

When hab pkg download is provided a fully-qualified identifier, it
will fail if that package isn't in the stable channel. In other words
we would need to provide a channel that contains the package on the
command line.  That's a departure from other CLI commands since a
fully-qualified identifier is enough to find a package regardless of
what channel it might be in.

Signed-off-by: Mark Anderson <mark@chef.io>