Initialize health check gateway state to Unknown
Previously, the gateway state had no health check information for a
given service until the first health-check. This resulted in a 404 for
the `/services/SERVICE_NAME/SERVICE_GROUP/health` endpoint. However,
according to the documentation, a 404 response code means the service
isn't loaded:

```
 /{name}/{group}/health:
     get:
         description: Health check status and output for the given service group
         responses:
             200:
                 description: Health Check - Ok / Warning
                 body:
                     application/json:
                         type: healthCheckOutput
             404:
                 description: Service not loaded
             500:
                 description: Health Check - Unknown
             503:
                 description: Health Check - Critical
```

This change avoids the problem by populating the gateway state with
Unknown when we start the health checks. As a result we now return 500
for the health endpoint of a loaded service until the first
health-check comes back, in accordance with the API documentation.

Fixes #6747

Signed-off-by: Steven Danna <steve@chef.io>