[CI] Make Supervisor connection failure test more robust
We were seeing failures for this test in CI, but not locally. I
hypothesize that this is because the test was waiting 5 seconds after
starting a Supervisor to observe the desired state. However, the bulk
of that time was actually occupied by downloading and installing
`core/hab-sup` and `core/hab-launcher` packages, which are not present
on the clean systems we test with. As a result, we were failing the
test unnecessarily.

Here, we preinstall these packages before starting the Supervisor and
entering our waiting period.

Additionally, we make some additional assertions about the nature of
the observed error by comparing the Supervisor's output with a desired
string. While this does couple the test to log output, it provides us
a bit more certainty that the test is actually testing what we want it
to, and not "passing" for some unexpected reason. The error message in
the event that this match no longer works should help guide developers
to a suitable fix quickly, should the need arise.

Signed-off-by: Christopher Maier <cmaier@chef.io>